UPDATE/Repo_Service

+++ KoiFarmShop/KoiFarmShop.Service/IServices/IKoiFishService.cs
@@ -5,7 +5,7 @@ namespace KoiFarmShop.Service;
 public interface IKoiFishService
 {
     Task<IEnumerable<KoiFish>> GetAllKoiFishAsync();
-    Task<KoiFish> GetKoiFishByIdAsync(int id);
+    Task<KoiFish> GetKoiFishByIdAsync(long id);
     Task AddKoiFishAsync(KoiFish koiFish);
     Task UpdateKoiFishAsync(KoiFish koiFish);
     Task DeleteKoiFishAsync(int id);
